rndom valid

+++ src/main/java/pl/agh/kis/seirsimulation/model/strategy/DiseaseStrategy.java
@@ -5,8 +5,8 @@ import pl.agh.kis.seirsimulation.model.Cell;
 import java.util.List;
 
 public interface DiseaseStrategy {
-    int calculateSusceptibleChange(List<Integer> stateCount,int stateCountSum);
-    int calculateExposedChange(List<Integer> stateCount,int stateCountSum);
+    int calculateSusceptibleChange(List<Integer> stateCount,double stateCountSum);
+    int calculateExposedChange(List<Integer> stateCount,double stateCountSum);
     int calculateInfectedChange(List<Integer> stateCount);
     int calculateRecoveredChange(List<Integer> stateCount);
     int calculateDiseaseDeaths(List<Integer> stateCount);
